refactor(services): split handlers into separate modules
